Split tests into separate files
`JSONSchemerTest` was getting too long.
